feat: Restructured the app files and folder.
